+++ src/client/scripts/theme.ts
@@ -9,22 +9,22 @@ export type Theme = {
 	props: { [key: string]: string };
 };
 
-export const lightTheme: Theme = require('./themes/_light.json5');
-export const darkTheme: Theme = require('./themes/_dark.json5');
+export const lightTheme: Theme = require('../themes/_light.json5');
+export const darkTheme: Theme = require('../themes/_dark.json5');
 
 export const builtinThemes = [
-	require('./themes/white.json5'),
-	require('./themes/black.json5'),
-	require('./themes/lavender.json5'),
-	require('./themes/halloween.json5'),
-	require('./themes/city.json5'),
-	require('./themes/rainy.json5'),
-	require('./themes/urban.json5'),
-	require('./themes/cafe.json5'),
-	require('./themes/chocolate.json5'),
-	require('./themes/danboard.json5'),
-	require('./themes/olive.json5'),
-	require('./themes/ocean.json5'),
+	require('../themes/white.json5'),
+	require('../themes/black.json5'),
+	require('../themes/lilac.json5'),
+	require('../themes/halloween.json5'),
+	require('../themes/city.json5'),
+	require('../themes/rainy.json5'),
+	require('../themes/urban.json5'),
+	require('../themes/cafe.json5'),
+	require('../themes/chocolate.json5'),
+	require('../themes/danboard.json5'),
+	require('../themes/olive.json5'),
+	require('../themes/ocean.json5'),
 ] as Theme[];
 
 let timeout = null;
