+++ src/client/themes/lilac.json5
@@ -1,7 +1,7 @@
 {
 	id: 'e9c8c01d-9c15-48d0-9b5c-3d00843b5b36',
 
-	name: 'Lavender',
+	name: 'Lilac',
 	author: 'syuilo',
 
 	base: 'light',
